+++ src/test/java/io/xpring/xrpl/helpers/XrpTestUtils.java
@@ -1,13 +1,17 @@
 package io.xpring.xrpl.helpers;
 
 import io.xpring.common.XrplNetwork;
+import io.xpring.xrpl.model.MemoField;
+import io.xpring.xrpl.model.XrpMemo;
 import io.xpring.xrpl.model.XrpTransaction;
 import org.xrpl.rpc.v1.GetAccountTransactionHistoryResponse;
 import org.xrpl.rpc.v1.GetTransactionResponse;
+import org.xrpl.rpc.v1.Memo;
 import org.xrpl.rpc.v1.Transaction;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 /**
  * Convenience class for utility functions used in test cases for XrpClient infrastructure.
@@ -43,4 +47,41 @@ public class XrpTestUtils {
     }
     return payments;
   }
+
+  private static Optional<MemoField> memoField1 = Optional.of(
+          MemoField.builder().value("I forgot to pick up Carl...").isHex(false).build()
+  );
+  private static Optional<MemoField> memoField2 = Optional.of(
+          MemoField.builder().value("jaypeg").isHex(false).build()
+  );
+  private static Optional<MemoField> memoField3 = Optional.of(
+          MemoField.builder().value("meme").isHex(false).build()
+  );
+
+  private static Optional<MemoField> blankMemoField = Optional.of(
+          MemoField.builder().value("").isHex(false).build()
+  );
+
+  public static XrpMemo iForgotToPickUpCarlMemo = XrpMemo.fromMemoFields(memoField1, memoField2, memoField3);
+
+  public static XrpMemo noDataMemo = XrpMemo.fromMemoFields(Optional.empty(), memoField2, memoField3);
+
+  /**
+   * Exists because ledger will store value as blank.
+   */
+  public static XrpMemo expectedNoDataMemo = XrpMemo.fromMemoFields(blankMemoField, memoField2, memoField3);
+
+  public static XrpMemo noFormatMemo = XrpMemo.fromMemoFields(memoField1, Optional.empty(), memoField3);
+
+  /**
+   * Exists because ledger will stored value as blank.
+   */
+  public static XrpMemo expectedNoFormatMemo = XrpMemo.fromMemoFields(memoField1, blankMemoField, memoField3);
+
+  public static XrpMemo noTypeMemo = XrpMemo.fromMemoFields(memoField1, memoField2, Optional.empty());
+
+  /**
+   * Exists because ledger will stored value as blank.
+   */
+  public static XrpMemo expectedNoTypeMemo = XrpMemo.fromMemoFields(memoField1, memoField2, blankMemoField);
 }